+++ apps/session-manager/src/app.ts
@@ -78,6 +78,7 @@ import { bearerWalletTokenStrategy } from "./auth/bearer-wallet-token-strategy";
 import { AcsDependencies } from "./controllers/authentication";
 import { localStrategy } from "./auth/local-strategy";
 import { FF_LOLLIPOP_ENABLED } from "./config/lollipop";
+import { getServerUnavailableMiddleware } from "./utils/middlewares/server-load";
 
 export interface IAppFactoryParameters {
   readonly appInsightsClient?: appInsights.TelemetryClient;
@@ -145,15 +146,7 @@ export const newApp: (
   const ZENDESK_BASE_PATH = getRequiredENVVar("ZENDESK_BASE_PATH");
 
   // Server Unavailable Middleware
-  app.use((_req, res, next) => {
-    const isServerUnderPressure = app.get("isServerUnderPressure") as boolean;
-
-    if (isServerUnderPressure) {
-      res.status(503).send();
-    } else {
-      next();
-    }
-  });
+  app.use(getServerUnavailableMiddleware(app));
 
   // Setup paths
